Build the basic instrument (80%)
now the generator is ready, as long as the "isValid" funciton implemented by recursive dfs
add a consoleapp project to easily debug in console mod when the qt part is not finished yet

+++ RowVisitor.cpp
@@ -10,11 +10,11 @@ namespace MazeCracker
 			// 把指针和行号放进去即可
 		}
 		
-		const ICell& RowVisitor::operator[](const size_t& index) const
+		const ICell& RowVisitor::operator[](const int& index) const
 		{
 			return (*const_cast<IMaze const*>(implementer))[row][index];
 		}
-		ICell& RowVisitor::operator[](const size_t& index)
+		ICell& RowVisitor::operator[](const int& index)
 		{
 			return implementer->getCell(row, index);
 		}
